Send Kalman Values to Pc
Send Values as float instead of string

+++ Core/Inc/mpu6050.h
@@ -33,8 +33,8 @@ typedef struct
 
     float Temperature;
 
-    double KalmanAngleX;
-    double KalmanAngleY;
+    float KalmanAngleX;
+    float KalmanAngleY;
 } MPU6050_t;
 
 // Kalman structure
